config jpa instead of mybatis

+++ src/main/java/com/github/ksewen/ganyu/security/Authentication.java
@@ -2,7 +2,8 @@ package com.github.ksewen.ganyu.security;
 
 public interface Authentication {
 
-    Integer getUserId();
+    Long getUserId();
 
     String getUsername();
+
 }
